working app add delete edit products with Modal pop up for Admin only

+++ store/actions/products.js
@@ -37,7 +37,9 @@ export const addProduct = ( productData ) => {
             .then( response => {
               //  console.log('productData', productData);
                 dispatch( addProductSuccess( response.data.name, productData  ) );
-                Router.push('/products');
+              //  Router.push('/products');
+              dispatch( addClose());
+              dispatch(fetchProducts());
             } )
             .catch( error => {
                 console.log(error);
@@ -63,11 +65,7 @@ export const fetchProductsStart = () => {
 };
 
 export const fetchProductsSuccess = (productsArr) => {
-    // console.log('array?:',productsArr);
-    // if (!Array.isArray(productsArr)){
-    //     productsArr=[...productsArr];
-    //     console.log('not array:',productsArr);
-    // } else { console.log('productsArr is array');}
+    // orderBy product_name
     return {
         type: actionTypes.FETCH_PRODUCTS_SUCCESS,
         loading: false,
@@ -82,13 +80,14 @@ export const fetchProducts = () => {
     axios.get( '/products.json' )
         .then( res => {
                 const fetchedProducts = [];
+                console.log('res.data',res.data);
                 for ( let key in res.data ) {
                     let product = {};
                     product.id = key;
                     product.productData = res.data[key].productData
                     fetchedProducts.push( product );
                 }
-                //console.log('fetchedProducts',[...fetchedProducts]);
+                console.log('fetchedProducts',[...fetchedProducts]);
                 dispatch(fetchProductsSuccess([...fetchedProducts]));
             })
              .catch( err => {
@@ -111,24 +110,42 @@ export const deleteProductStart = () => {
 
 export const deleteProduct = (id) => {
     return dispatch => {
-        dispatch(deleteProductStart());
-    axios.delete( '/products/' + id + '.json', null )
-        .then( res => {
-                console.log('delete product success', res);
-                dispatch(fetchProducts());
-            })
-        .catch( err => {
-                console.log('delete product fail', err);
-                // dispatch(deleteProductsFail(err));
-             } );  
-};
+        if (id) {
+            dispatch(deleteProductStart());
+            axios.delete('/products/' + id + '.json', null)
+                .then(res => {
+                    console.log('delete product success', res);
+                    dispatch(fetchProducts());
+                })
+                .catch(err => {
+                    console.log('delete product fail', err);
+                    // dispatch(deleteProductsFail(err));
+                });
+        } else {
+            console.warn('no id - cannot delete all!');
+        }
+    };
 };
 
+export const addOpen = () => {
+    return {
+        type: actionTypes.ADD_OPEN,
+        addShow: true,
+    };
+};
 
-export const editOpen = () => {
+export const addClose = () => {
+    return {
+        type: actionTypes.ADD_CLOSE,
+        addShow: false
+    };
+};
+
+export const editOpen = (pIn) => {
     return {
         type: actionTypes.EDIT_OPEN,
-        editShow: true
+        editShow: true,
+        p_in: pIn
     };
 };
 
@@ -136,5 +153,47 @@ export const editClose = () => {
     return {
         type: actionTypes.EDIT_CLOSE,
         editShow: false
+    };
+};
+
+
+export const editProductStart = () => {
+    return {
+        type: actionTypes.EDIT_PRODUCT_START
+    };
+};
+
+export const editProductSuccess = () => {
+    return {
+        type: actionTypes.EDIT_PRODUCT_SUCCESS,
+        loading: false
+    };
+};
+
+export const editProductFail = ( error ) => {
+    return {
+        type: actionTypes.EDIT_PRODUCT_FAIL,
+        error: error
+    };
+};
+
+export const editProduct = (id, productData ) => {
+    console.log('action id', id);
+    console.log('action productData', productData);
+    return dispatch => {
+       dispatch( editProductStart() );
+      //  ?auth=' + token + token,  rest/saving-data/products.json
+         axios.put( '/products/'+ id +'/productData.json', productData )
+            .then( response => {
+                console.log('action Editproduct Success', response);
+                dispatch( editProductSuccess( ) );
+              
+                dispatch( editClose());
+                dispatch(fetchProducts());
+            } )
+            .catch( error => {
+                console.log('axios edit error' + error);
+                dispatch( editProductFail( error ) );
+            } ); 
     };
 };
